+++ crates/networking/p2p/sync.rs
@@ -1,18 +1,22 @@
+mod trie_rebuild;
+mod storage_fetcher;
+
+
 use ethrex_blockchain::error::ChainError;
 use ethrex_common::{
     types::{AccountState, Block, BlockHash, EMPTY_KECCACK_HASH},
     BigEndianHash, H256, U256, U512,
 };
 use ethrex_rlp::{decode::RLPDecode, encode::RLPEncode, error::RLPDecodeError};
-use ethrex_storage::{error::StoreError, Store, MAX_SNAPSHOT_READS, STATE_TRIE_SEGMENTS};
+use ethrex_storage::{error::StoreError, Store, STATE_TRIE_SEGMENTS};
 use ethrex_trie::{Nibbles, Node, TrieError, TrieState, EMPTY_TRIE_HASH};
+use trie_rebuild::TrieRebuilder;
 use std::{array, cmp::min, collections::BTreeMap, sync::Arc};
 use tokio::{
     sync::{
         mpsc::{self, error::SendError, Receiver, Sender},
         Mutex,
     },
-    task::JoinSet,
     time::{sleep, Duration, Instant},
 };
 use tokio_util::sync::CancellationToken;
@@ -70,52 +74,6 @@ pub struct SyncManager {
     trie_rebuilder: Option<TrieRebuilder>,
     // Used for cancelling long-living tasks upon shutdown
     cancel_token: CancellationToken,
-}
-
-/// Represents the permanently ongoing background trie rebuild process
-/// This process will be started whenever a state sync is initiated and will be
-/// kept alive throughout sync cycles, only stopping once the tries are fully rebuilt or the node is stopped
-#[derive(Debug)]
-struct TrieRebuilder {
-    state_trie_rebuilder: tokio::task::JoinHandle<Result<(), SyncError>>,
-    storage_trie_rebuilder: tokio::task::JoinHandle<Result<(), SyncError>>,
-    storage_rebuilder_sender: Sender<Vec<(H256, H256)>>,
-}
-
-impl TrieRebuilder {
-    /// Returns true is the trie rebuild porcess is alive and well
-    fn alive(&self) -> bool {
-        !(self.state_trie_rebuilder.is_finished()
-            || self.storage_trie_rebuilder.is_finished()
-            || self.storage_rebuilder_sender.is_closed())
-    }
-    /// Waits for the rebuild process to complete and returns the resulting mismatched accounts
-    async fn complete(self) -> Result<(), SyncError> {
-        // Signal storage rebuilder to finish
-        self.storage_rebuilder_sender.send(vec![]).await?;
-        self.state_trie_rebuilder.await??;
-        self.storage_trie_rebuilder.await?
-    }
-
-    /// starts the background trie rebuild process
-    fn startup(cancel_token: CancellationToken, store: Store) -> Self {
-        let (storage_rebuilder_sender, storage_rebuilder_receiver) =
-            mpsc::channel::<Vec<(H256, H256)>>(MAX_CHANNEL_MESSAGES);
-        let state_trie_rebuilder = tokio::task::spawn(rebuild_state_trie_in_backgound(
-            store.clone(),
-            cancel_token.clone(),
-        ));
-        let storage_trie_rebuilder = tokio::task::spawn(rebuild_storage_trie_in_background(
-            store,
-            cancel_token,
-            storage_rebuilder_receiver,
-        ));
-        Self {
-            state_trie_rebuilder,
-            storage_trie_rebuilder,
-            storage_rebuilder_sender,
-        }
-    }
 }
 
 impl SyncManager {
@@ -710,206 +668,6 @@ async fn fetch_bytecode_batch(
     Ok(batch)
 }
 
-/// Waits for incoming account hashes & storage roots from the receiver channel endpoint, queues them, and fetches and stores their bytecodes in batches
-/// This function will remain active until either an empty vec is sent to the receiver or the pivot becomes stale
-/// Upon finsih, remaining storages will be sent to the storage healer
-async fn storage_fetcher(
-    mut receiver: Receiver<Vec<(H256, H256)>>,
-    peers: PeerHandler,
-    store: Store,
-    state_root: H256,
-    storage_trie_rebuilder_sender: Sender<Vec<(H256, H256)>>,
-    storage_healer_sender: Sender<Vec<H256>>,
-) -> Result<(), SyncError> {
-    // Pending list of storages to fetch
-    let mut pending_storage: Vec<(H256, H256)> = vec![];
-    // The pivot may become stale while the fetcher is active, we will still keep the process
-    // alive until the end signal so we don't lose queued messages
-    let mut stale = false;
-    let mut incoming = true;
-    while incoming {
-        // Fetch incoming requests
-        let mut msg_buffer = vec![];
-        if receiver.recv_many(&mut msg_buffer, MAX_CHANNEL_READS).await != 0 {
-            for account_hashes_and_roots in msg_buffer {
-                if !account_hashes_and_roots.is_empty() {
-                    pending_storage.extend(account_hashes_and_roots);
-                } else {
-                    // Empty message signaling no more bytecodes to sync
-                    incoming = false
-                }
-            }
-        } else {
-            // Disconnect
-            incoming = false
-        }
-        // If we have enough pending bytecodes to fill a batch
-        // or if we have no more incoming batches, spawn a fetch process
-        // If the pivot became stale don't process anything and just save incoming requests
-        while !stale
-            && (pending_storage.len() >= BATCH_SIZE || (!incoming && !pending_storage.is_empty()))
-        {
-            // We will be spawning multiple tasks and then collecting their results
-            // This uses a loop inside the main loop as the result from these tasks may lead to more values in queue
-            let mut storage_tasks = tokio::task::JoinSet::new();
-            for _ in 0..MAX_PARALLEL_FETCHES {
-                let next_batch = pending_storage
-                    .drain(..BATCH_SIZE.min(pending_storage.len()))
-                    .collect::<Vec<_>>();
-                storage_tasks.spawn(fetch_storage_batch(
-                    next_batch,
-                    state_root,
-                    peers.clone(),
-                    store.clone(),
-                    storage_trie_rebuilder_sender.clone(),
-                ));
-                // End loop if we don't have enough elements to fill up a batch
-                if pending_storage.is_empty() || (incoming && pending_storage.len() < BATCH_SIZE) {
-                    break;
-                }
-            }
-            // Add unfetched accounts to queue and handle stale signal
-            for res in storage_tasks.join_all().await {
-                let (remaining, is_stale) = res?;
-                pending_storage.extend(remaining);
-                stale |= is_stale;
-            }
-        }
-    }
-    debug!(
-        "Concluding storage fetcher, {} storages left in queue to be healed later",
-        pending_storage.len()
-    );
-    if !pending_storage.is_empty() {
-        storage_healer_sender
-            .send(pending_storage.into_iter().map(|(hash, _)| hash).collect())
-            .await?;
-    }
-    Ok(())
-}
-
-/// Receives a batch of account hashes with their storage roots, fetches their respective storage ranges via p2p and returns a list of the code hashes that couldn't be fetched in the request (if applicable)
-/// Also returns a boolean indicating if the pivot became stale during the request
-async fn fetch_storage_batch(
-    mut batch: Vec<(H256, H256)>,
-    state_root: H256,
-    peers: PeerHandler,
-    store: Store,
-    storage_trie_rebuilder_sender: Sender<Vec<(H256, H256)>>,
-) -> Result<(Vec<(H256, H256)>, bool), SyncError> {
-    // A list of all completely fetched storages to send to the rebuilder
-    let mut complete_storages = vec![];
-    debug!(
-        "Requesting storage ranges for addresses {}..{}",
-        batch.first().unwrap().0,
-        batch.last().unwrap().0
-    );
-    let (batch_hahses, batch_roots) = batch.clone().into_iter().unzip();
-    if let Some((mut keys, mut values, incomplete)) = peers
-        .request_storage_ranges(state_root, batch_roots, batch_hahses, H256::zero())
-        .await
-    {
-        debug!("Received {} storage ranges", keys.len(),);
-        // Handle incomplete ranges
-        if incomplete {
-            // An incomplete range cannot be empty
-            let (last_keys, last_values) = (keys.pop().unwrap(), values.pop().unwrap());
-            // If only one incomplete range is returned then it must belong to a trie that is too big to fit into one request
-            // We will handle this large trie separately
-            if keys.is_empty() {
-                debug!("Large storage trie encountered, handling separately");
-                let (account_hash, storage_root) = batch.remove(0);
-                if handle_large_storage_range(
-                    state_root,
-                    account_hash,
-                    storage_root,
-                    last_keys,
-                    last_values,
-                    peers.clone(),
-                    store.clone(),
-                )
-                .await?
-                {
-                    // Pivot became stale
-                    // Add trie back to the queue and return stale pivot status
-                    batch.push((account_hash, storage_root));
-                    return Ok((batch, true));
-                }
-                // Add large storage to completed storages
-                complete_storages.push((account_hash, storage_root));
-            }
-            // The incomplete range is not the first, we cannot asume it is a large trie, so lets add it back to the queue
-        }
-        // Store the storage ranges & rebuild the storage trie for each account
-        for (keys, values) in keys.into_iter().zip(values.into_iter()) {
-            let (account_hash, expected_root) = batch.remove(0);
-            // Write storage to snapshot
-            store.write_snapshot_storage_batch(account_hash, keys, values)?;
-            complete_storages.push((account_hash, expected_root));
-        }
-        // Send complete storages to the rebuilder
-        storage_trie_rebuilder_sender
-            .send(complete_storages)
-            .await?;
-        // Return remaining code hashes in the batch if we couldn't fetch all of them
-        return Ok((batch, false));
-    }
-    // Pivot became stale
-    Ok((batch, true))
-}
-
-/// Handles the returned incomplete storage range of a large storage trie and
-/// fetches the rest of the trie using single requests
-/// Returns a boolean indicating is the pivot became stale during fetching
-// TODO: Later on this method can be refactored to use a separate queue process
-// instead of blocking the current thread for the remainder of the retrieval
-async fn handle_large_storage_range(
-    state_root: H256,
-    account_hash: H256,
-    storage_root: H256,
-    keys: Vec<H256>,
-    values: Vec<U256>,
-    peers: PeerHandler,
-    store: Store,
-) -> Result<bool, SyncError> {
-    // First process the initial range
-    // Keep hold of the last key as this will be the first key of the next range
-    let mut next_key = *keys.last().unwrap();
-    let mut current_root = {
-        let mut trie = store.open_storage_trie(account_hash, *EMPTY_TRIE_HASH);
-        for (key, value) in keys.into_iter().zip(values.into_iter()) {
-            trie.insert(key.0.to_vec(), value.encode_to_vec())?;
-        }
-        // Compute current root so we can extend this trie later
-        trie.hash()?
-    };
-    let mut should_continue = true;
-    // Fetch the remaining range
-    while should_continue {
-        debug!("Fetching large storage trie, current key: {}", next_key);
-
-        if let Some((keys, values, incomplete)) = peers
-            .request_storage_range(state_root, storage_root, account_hash, next_key)
-            .await
-        {
-            next_key = *keys.last().unwrap();
-            should_continue = incomplete;
-            let mut trie = store.open_storage_trie(account_hash, current_root);
-            for (key, value) in keys.into_iter().zip(values.into_iter()) {
-                trie.insert(key.0.to_vec(), value.encode_to_vec())?;
-            }
-            // Compute current root so we can extend this trie later
-            current_root = trie.hash()?;
-        } else {
-            return Ok(true);
-        }
-    }
-    if current_root != storage_root {
-        warn!("State sync failed for storage root {storage_root}");
-    }
-    Ok(false)
-}
-
 /// Heals the trie given its state_root by fetching any missing nodes in it via p2p
 /// Returns true if healing was fully completed or false if we need to resume healing on the next sync cycle
 async fn heal_state_trie(
@@ -1178,232 +936,6 @@ fn node_missing_children(
     }
     Ok(paths)
 }
-
-#[derive(Debug, Clone)]
-pub(crate) struct SegmentStatus {
-    pub current: H256,
-    pub end: H256,
-}
-
-impl SegmentStatus {
-    pub(crate) fn complete(&self) -> bool {
-        self.current >= self.end
-    }
-}
-
-/// Rebuilds the state trie by processing the accounts from the state snapshot
-/// Will only stop when state sync has finished and all account have been processed or when the cancel token is cancelled
-async fn rebuild_state_trie_in_backgound(
-    store: Store,
-    cancel_token: CancellationToken,
-) -> Result<(), SyncError> {
-    // Get initial status from checkpoint if available (aka node restart)
-    let checkpoint = store.get_state_trie_rebuild_checkpoint()?;
-    let mut rebuild_status = array::from_fn(|i| SegmentStatus {
-        current: checkpoint
-            .map(|(_, ch)| ch[i])
-            .unwrap_or(STATE_TRIE_SEGMENTS_START[i]),
-        end: STATE_TRIE_SEGMENTS_END[i],
-    });
-    let mut root = checkpoint.map(|(root, _)| root).unwrap_or(*EMPTY_TRIE_HASH);
-    let mut current_segment = 0;
-    let start_time = Instant::now();
-    let initial_rebuild_status = rebuild_status.clone();
-    let mut last_show_progress = Instant::now();
-    while !rebuild_status.iter().all(|status| status.complete()) {
-        // Show Progress stats (this task is not vital so we can detach it)
-        if Instant::now().duration_since(last_show_progress) >= SHOW_PROGRESS_INTERVAL_DURATION {
-            last_show_progress = Instant::now();
-            tokio::spawn(show_trie_rebuild_progress(
-                start_time,
-                initial_rebuild_status.clone(),
-                rebuild_status.clone(),
-            ));
-        }
-        // Check for cancellation signal from the main node execution
-        if cancel_token.is_cancelled() {
-            return Ok(());
-        }
-        if !rebuild_status[current_segment].complete() {
-            // Start rebuilding the current trie segment
-            let (current_root, current_hash) = rebuild_state_trie_segment(
-                root,
-                rebuild_status[current_segment].current,
-                current_segment,
-                store.clone(),
-                cancel_token.clone(),
-            )
-            .await?;
-            // Update status
-            root = current_root;
-            rebuild_status[current_segment].current = current_hash;
-        }
-        // Update DB checkpoint
-        let checkpoint = (root, rebuild_status.clone().map(|st| st.current));
-        store.set_state_trie_rebuild_checkpoint(checkpoint)?;
-        // Move on to the next segment
-        current_segment = (current_segment + 1) % STATE_TRIE_SEGMENTS
-    }
-
-    Ok(())
-}
-
-/// Fetches accounts from the state snasphot starting from the `start` hash and adds them to the trie
-/// Will stop when there are no more accounts within the segment bounds in the snapshot, or when the cancel token is cancelled
-// Returns the current root, the last processed account hash
-// If state sync is finished and there are no more snapshot accounts for the segment the account hash
-// returned will be the segment bound to notify that the segment has been fully rebuilt
-async fn rebuild_state_trie_segment(
-    mut root: H256,
-    mut start: H256,
-    segment_number: usize,
-    store: Store,
-    cancel_token: CancellationToken,
-) -> Result<(H256, H256), SyncError> {
-    let mut state_trie = store.open_state_trie(root);
-    loop {
-        if cancel_token.is_cancelled() {
-            break;
-        }
-        let mut batch = store.read_account_snapshot(start)?;
-        // Remove out of bounds elements
-        batch.retain(|(hash, _)| *hash <= STATE_TRIE_SEGMENTS_END[segment_number]);
-        let unfilled_batch = batch.len() < MAX_SNAPSHOT_READS;
-        // Update start
-        if let Some(last) = batch.last() {
-            start = next_hash(last.0);
-        }
-        // Process batch
-        // Add accounts to the state trie
-        for (hash, account) in batch.iter() {
-            state_trie.insert(hash.0.to_vec(), account.encode_to_vec())?;
-        }
-        root = state_trie.hash()?;
-        // Return if we have no more snapshot accounts to process for this segemnt
-        if unfilled_batch {
-            let state_sync_complete = store
-                .get_state_trie_key_checkpoint()?
-                .is_some_and(|ch| ch[segment_number] == STATE_TRIE_SEGMENTS_END[segment_number]);
-            // Mark segment as finished if state sync is complete
-            if state_sync_complete {
-                start = STATE_TRIE_SEGMENTS_END[segment_number];
-            }
-            break;
-        }
-    }
-    Ok((root, start))
-}
-
-/// Waits for incoming messages from the storage fetcher and rebuilds the associated storages
-/// Will stop when the stop signal is received (an empty vec) and there are no more storages in queue or when the cancel token is cancelled
-// Only receives fully downloaded storages, and will only emit a warning if there is a mismatch between the expected root and the rebuilt root, as this is considered a bug
-async fn rebuild_storage_trie_in_background(
-    store: Store,
-    cancel_token: CancellationToken,
-    mut receiver: Receiver<Vec<(H256, H256)>>,
-) -> Result<(), SyncError> {
-    // (AccountHash, ExpectedRoot)
-    let mut pending_storages = store
-        .get_storage_trie_rebuild_pending()?
-        .unwrap_or_default();
-    let mut incoming = true;
-    while incoming || !pending_storages.is_empty() {
-        if cancel_token.is_cancelled() {
-            break;
-        }
-        // Read incoming batch
-        if !receiver.is_empty() || pending_storages.is_empty() {
-            let mut buffer = vec![];
-            receiver.recv_many(&mut buffer, MAX_CHANNEL_READS).await;
-            incoming = !buffer.iter().any(|batch| batch.is_empty());
-            pending_storages.extend(buffer.iter().flatten());
-        }
-
-        // Spawn tasks to rebuild current storages
-        let mut rebuild_tasks = JoinSet::new();
-        for _ in 0..MAX_PARALLEL_FETCHES {
-            if pending_storages.is_empty() {
-                break;
-            }
-            let (account_hash, expected_root) = pending_storages.pop().unwrap();
-            let store = store.clone();
-            rebuild_tasks.spawn(rebuild_storage_trie(
-                account_hash,
-                expected_root,
-                store.clone(),
-            ));
-        }
-        for res in rebuild_tasks.join_all().await {
-            res?;
-        }
-    }
-    store.set_storage_trie_rebuild_pending(pending_storages)?;
-    Ok(())
-}
-
-/// Rebuilds a storage trie by reading from the storage snapshot
-/// Assumes that the storage has been fully downloaded and will only emit a warning if there is a mismatch between the expected root and the rebuilt root, as this is considered a bug
-async fn rebuild_storage_trie(
-    account_hash: H256,
-    expected_root: H256,
-    store: Store,
-) -> Result<(), SyncError> {
-    let mut start = H256::zero();
-    let mut storage_trie = store.open_storage_trie(account_hash, *EMPTY_TRIE_HASH);
-    loop {
-        let batch = store.read_storage_snapshot(account_hash, start)?;
-        let unfilled_batch = batch.len() < MAX_SNAPSHOT_READS;
-        // Update start
-        if let Some(last) = batch.last() {
-            start = next_hash(last.0);
-        }
-        // Process batch
-        // Launch storage rebuild tasks for all non-empty storages
-        for (key, val) in batch {
-            storage_trie.insert(key.0.to_vec(), val.encode_to_vec())?;
-        }
-        storage_trie.hash()?;
-
-        // Return if we have no more snapshot accounts to process for this segemnt
-        if unfilled_batch {
-            break;
-        }
-    }
-    if storage_trie.hash()? != expected_root {
-        warn!("Mismatched storage root for account {account_hash}");
-    }
-    Ok(())
-}
-
-async fn show_trie_rebuild_progress(
-    start_time: Instant,
-    initial_rebuild_status: [SegmentStatus; STATE_TRIE_SEGMENTS],
-    rebuild_status: [SegmentStatus; STATE_TRIE_SEGMENTS],
-) {
-    // Count how many hashes we already inserted in the trie and how many we inserted this cycle
-    let mut accounts_processed = U256::zero();
-    let mut accounts_processed_this_cycle = U256::one();
-    for i in 0..STATE_TRIE_SEGMENTS {
-        accounts_processed +=
-            rebuild_status[i].current.into_uint() - (STATE_TRIE_SEGMENTS_START[i].into_uint());
-        accounts_processed_this_cycle +=
-            rebuild_status[i].current.into_uint() - initial_rebuild_status[i].current.into_uint();
-    }
-    // Calculate completion rate
-    let completion_rate = (U512::from(accounts_processed + U256::one()) * U512::from(100))
-        / U512::from(U256::max_value());
-    // Time to finish = Time since start / Accounts processed this cycle * Remaining accounts
-    let remaining_accounts = U256::MAX - accounts_processed;
-    let time_to_finish = (U512::from(start_time.elapsed().as_secs())
-        * U512::from(remaining_accounts))
-        / (U512::from(accounts_processed_this_cycle));
-    info!(
-        "State Trie Rebuild Progress: {}%, estimated time to finish: {}",
-        completion_rate,
-        seconds_to_readable(time_to_finish)
-    );
-}
-
 #[derive(Clone)]
 struct StateSyncProgress {
     data: Arc<Mutex<StateSyncProgressData>>,
@@ -1499,11 +1031,6 @@ fn seconds_to_readable(seconds: U512) -> String {
         return format!("Over {days} days");
     }
     format!("{hours}h{minutes}m{seconds}s")
-}
-
-/// Returns hash + 1
-fn next_hash(hash: H256) -> H256 {
-    H256::from_uint(&(hash.into_uint() + 1))
 }
 
 #[derive(thiserror::Error, Debug)]
